Create seperate file for EntityShareAction and adjust EntitySHareStore default export
