+++ graylog2-web-interface/src/stores/users/UsersStore.js
@@ -266,5 +266,4 @@ const UsersStore: UsersStoreType = singletonStore(
 );
 
 export { UsersActions, UsersStore };
-
 export default UsersStore;
